requirements.txt
langchain-core>=0.1.35
langchain-openai>=0.1.2
langgraph>=0.0.33
pydantic>=2.0
typing-extensions>=4.5.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.32.0
